Sales Number updated Correctly
